Updating sectioned form to use the tab select callback instead of $scope.$watch on the active tab
There was also a side effect of watching the active tab in that when the
form was loaded in mobile the active tab was changed, thus hiding the
tab content by default.  The mediaBreakpointHandler does not work in
this case because it is first run (on registration) in $onInit which has
not loaded child elements.  To overcome this added a $postLink function
that sets the initial form state after child elements are loaded.